analysis: Make field end offsets consistently exclusive
Summary:
Ghidra's DataTypeComponent.getEndOffset() is *inclusive*, e.g. [66, 66]
instead of [66, 67).  We were fixing this for bitfields but not regular
fields.  That caused us to miss some references to 1-byte fields, since
Field.atOffset() compares with `<`, not `<=`.

Fix it by always adding the length of the type itself to compute the
end offset.

Test Plan: ./scripts/analyze.sh bfs.csv bfs

Reviewers: ali, oscar

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D657