analysis/type: Try harder to de-duplicate types
Summary:
DWARF info for a type will be duplicated in every object file that uses
that type.  Ghidra tries to de-duplicate the type information, but fails
in some cases.  Particularly difficult are cases like

    foo.h:
    struct foo {
        struct bar *ptr;
    };

    bar.h:
    struct bar {
        int val;
    };

    foo.c:
    #include "foo.h"
    ...

    bar.c:
    #include "foo.h"
    #include "bar.h"
    ...

foo.c only has a forward declaration of struct bar, while bar.c has a
complete definition.  That will lead the debug info for struct foo to be
different between foo.o and bar.o, as only bar.o has the definition of
struct bar.

In order to de-duplicate struct foo in this case, we attempt to resolve
incomplete types to a complete type with the same name (potentially in a
different shared object) before wrapping them as a BcpiType.

We also look through typedefs in struct fields, as they are recorded
inconsistently.  Finally, we ignore the name and size of function types,
as those shouldn't matter for equivalence.

Test Plan:
./scripts/analyze.sh kernel kernel.csv
ls results/kernel/structs/tcpcb*

Reviewers: oscar, ali

Differential Revision: https://review.rcs.uwaterloo.ca/D900