analysis/cfg: Replace findSinksAndLoops() with a standard algorithm
Summary:
Our implementation of findSinksAndLoops() is equivalent to Ghidra's
getEntryPoints() on the reverse CFG, at least for nodes reachable from
the source.  Since all nodes are reachable from the virtual SOURCE node
at the point of the call, just use the standard algorithm.

This should also fix a crash on empty CFGs introduced in D891, because
getSuccessors() returns null (rather than an empty list) on nodes that
are not in the graph.

Test Plan: ./scripts/analyze.sh kernel kernel.csv

Reviewers: oscar, ali

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D901